Prepare for legacy tag importer: move createImporter to sub class
